Update factories files
