+++ edc-client-dotnet/factory/DocumentationItemFactory.cs
@@ -1,12 +1,19 @@
 ﻿using edc_client_dotnet.internalImpl.model;
+using edc_client_dotnet.model;
 
 namespace edc_client_dotnet.factory
 {
     public class DocumentationItemFactory
     {
-        public DocumentationItemService Create()
+        private readonly IServiceProvider _serviceProvider;
+
+        public DocumentationItemFactory(IServiceProvider serviceProvider)
         {
-            return new DocumentationItemService();
+            _serviceProvider = serviceProvider;
+        }
+        public IDocumentationItem Create()
+        {
+            return (IDocumentationItem)_serviceProvider.GetService(typeof(DocumentationItemImpl));
         }
     }
 }
