+++ edc-client-dotnet/factory/I18NFactory.cs
@@ -3,18 +3,18 @@ using edc_client_dotnet.model;
 
 namespace edc_client_dotnet.factory
 {
-    public class ContextItemFactory
+    public class I18NFactory
     {
         private readonly IServiceProvider _serviceProvider;
 
-        public ContextItemFactory(IServiceProvider serviceProvider)
+        public I18NFactory(IServiceProvider serviceProvider)
         {
             _serviceProvider = serviceProvider;
         }
 
-        public IContextItem Create()
+        public II18NContent Create()
         {
-            return (IContextItem)_serviceProvider.GetService(typeof(ContextItemImpl));
+            return (II18NContent)_serviceProvider.GetService(typeof(I18NContentImpl));
         }
     }
 }
